Move the cursor as you make selections
The rationale behind this is what if the line you want is nearly
a quarter screen above the split?  With the old way, you would need
to hit 'k' once, and then alternate between 'j' and 'k' to get to
the line you want, which seems like a waste of time.

This new change moves the cursor on each selection, so that if you
find that the split is close enough, you can just abort early.