Add option to start the split at the current line
This is an experiment - if this proves useful, it may end up
becoming the default behavior.

The rationale is that your eyes are used to being fixated on the
current line, and it's been a bit jarring for me to remember
"oh yeah, split navigate starts at the middle of the screen, not
the current line!".  This may increase the navigate time by a step
or two, but we'll see how things go.